hist mods

+++ reconstruction/plot_histogram.C
@@ -2,7 +2,7 @@ const int etabin = 3;
 static double eta_lo[etabin] = {-3,-1,1};
 static double eta_hi[etabin] = {-1,1,3};
 
-void plot_histogram(const char* inFile, const char* outDir, const char* title = "", const int events = 1000000)
+void plot_histogram(const char* inFile, const char* outDir, const char* title = "", const double events = 1000000)
 {
 
   TFile* f = new TFile(inFile,"read");
